@@ -11,19 +11,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/architect": "^0.1303.4",
-    "@angular-devkit/build-angular": "^13.3.6",
-    "@angular/animations": "^13.3.5",
-    "@angular/common": "^13.3.11",
-    "@angular/compiler": "^13.2.5",
-    "@angular/core": "^13.2.5",
-    "@angular/forms": "^13.3.11",
-    "@angular/localize": "^13.3.11",
-    "@angular/platform-browser": "^13.3.11",
-    "@angular/platform-browser-dynamic": "^13.3.11",
-    "@angular/router": "^13.3.11",
-    "@asymmetrik/ngx-leaflet": "^13.0.2",
-    "@ng-bootstrap/ng-bootstrap": "^12",
+    "@angular-devkit/architect": "^0.1401.0",
+    "@angular-devkit/build-angular": "^14.1.0",
+    "@angular/animations": "^14.1.0",
+    "@angular/common": "^14.1.0",
+    "@angular/compiler": "^14.1.0",
+    "@angular/core": "^14.1.0",
+    "@angular/forms": "^14.1.0",
+    "@angular/localize": "^14.1.0",
+    "@angular/platform-browser": "^14.1.0",
+    "@angular/platform-browser-dynamic": "^14.1.0",
+    "@angular/router": "^14.1.0",
+    "@asymmetrik/ngx-leaflet": "^14.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^13.0.0-beta.1",
     "@ng-bootstrap/schematics": "^2.0.0-alpha.1",
     "@ngx-grpc/common": "^3.0.0",
     "@ngx-grpc/core": "^3.0.1",
@@ -38,34 +38,34 @@
     "bootstrap": "^5",
     "chart.js": "^3.7.1",
     "core-js": "^3.22.7",
-    "keycloak-angular": "^10.0.0",
+    "keycloak-angular": "^12.0.0",
     "keycloak-js": "^18.0.0",
     "leaflet": "^1.8.0",
     "left-pad": "^1.3.0",
     "lodash.escaperegexp": "^4.1.2",
     "messageformat": "^2.3.0",
     "moment": "^2.29.3",
-    "ng-recaptcha": "^9.0.0",
-    "ng2-charts": "^3.0.10",
-    "ngx-autosize": "^1.8.4",
+    "ng-recaptcha": "^10.0.0",
+    "ng2-charts": "^4.0.0",
+    "ngx-autosize": "^2.0.2",
     "ngx-moment": "<6.0.0",
-    "ngx-pipes": "=3.0.0",
+    "ngx-pipes": "^3.1.0",
     "npm": "^8.3.1",
     "sprintf-js": "^1.1.2",
     "tslib": "^2.4.0",
-    "typescript": "4.5.5",
+    "typescript": "4.7.4",
     "url-parse": "^1.5.10",
     "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-eslint/builder": "^13.2.1",
-    "@angular-eslint/eslint-plugin": "^13.1.0",
-    "@angular-eslint/eslint-plugin-template": "^13.2.1",
-    "@angular-eslint/schematics": "^13.2.1",
-    "@angular-eslint/template-parser": "^13.2.1",
-    "@angular/cli": "^13.3.6",
-    "@angular/compiler-cli": "^13.3.5",
-    "@angular/language-service": "^13.3.11",
+    "@angular-eslint/builder": "^14.0.2",
+    "@angular-eslint/eslint-plugin": "^14.0.2",
+    "@angular-eslint/eslint-plugin-template": "^14.0.2",
+    "@angular-eslint/schematics": "^14.0.2",
+    "@angular-eslint/template-parser": "^14.0.2",
+    "@angular/cli": "^14.1.0",
+    "@angular/compiler-cli": "^14.1.0",
+    "@angular/language-service": "^14.1.0",
     "@brainly/onesky-utils": "^1.4.1",
     "@ngx-grpc/protoc-gen-ng": "^3.0.1",
     "@types/google-protobuf": "^3.15.6",
@@ -80,7 +80,7 @@
     "grpc-web": "^1.3.1",
     "html-comment-regex": "^1.1.2",
     "jquery": "^3.6.0",
-    "ngx-build-plus": "^13.0.1",
+    "ngx-build-plus": "^14.0.0",
     "showdown": "^2.1.0",
     "ts-node": "^10.7.0",
     "webpack": "^5.73.0",
