@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { LogComponent } from './log.component';
 
-const routes: Routes = [{ path: '', pathMatch: '', component: LogComponent }];
+const routes: Routes = [{ path: '', pathMatch: 'full', component: LogComponent }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
