@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { PolicyComponent } from './policy.component';
 
-const routes: Routes = [{ path: '', pathMatch: '', component: PolicyComponent }];
+const routes: Routes = [{ path: '', pathMatch: 'full', component: PolicyComponent }];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
